eval-in-emacs
For emacs users only. Enabling them to execute code in the slime
connected emacs from the lisp.

> (eval-in-emacs '(version))
"GNU Emacs 23.2.1 (x86_64-pc-linux-gnu, GTK+ Version 2.20.1)
 of 2010-12-11 on brahms, modified by Debian"

> (eval-in-emacs '(+ 1 2 3))
6

Useful for opening the editor on exising or new files.

> (eval-in-emacs '(progn (find-file "/tmp/foo123") nil))
NIL

Note that for EVAL-IN-EMACS to work you need to enable in
emacs, generally in the ~/.emacs file:

  (setq slime-enable-evaluate-in-emacs t)